requirements.txt
numpy==1.24.3
opencv-python==4.5.3.56
Pillow==9.5.0
kivy==2.1.0
kivymd==1.1.1